Use wikitext over Javascript where it's possible
Changing the media through Javascript was not appropriate (for example,
after playing a video, the player's controls did not appear when playing
a sound), and only worked with external files (_canonical_uri), not with
media tiddlers.

In the current solution, I use transclusion, which, in addition to
correcting the above-mentioned errors, also has the advantage of being
able to display ANY tiddler, so THE PLUGIN IS NOT JUST A MEDIA PLAYER
ANYMORE, BUT A GENERAL "SLIDESHOW MAKER". For example try to filter to
`[all[tiddlers+shadows]tag[$:/tags/Image]]`.

Jump button macros are pretty ugly, but I couldn't find a shorter form of
inserting the repeatr filter only when repeating is enabled.

WARNING: Commit introducing TODOs